sorc unlimited 1st circle spells
git-svn-id: svn+ssh://50.28.74.132/home/luminari/svn/luminari/myproject/trunk@413 dda7f814-ccd3-4591-a6d7-1bd23e5e64af

+++ memorize.c
@@ -521,7 +521,7 @@ int spellCircle(int class, int spellnum)
 {
   switch (class) {
     case CLASS_SORCERER:
-      return ((int)(MAX(1, (spell_info[spellnum].min_level[class]) / 2)));
+      return ((MAX(1, (spell_info[spellnum].min_level[class]) / 2)));
     /* can get confusing, just check out class.c to see what level
        they get their circles at in the spell_level function */
     case CLASS_PALADIN:
@@ -956,6 +956,7 @@ bool hasSpell(struct char_data *ch, int spellnum)
         return TRUE;
     }
   }
+  
 
   return FALSE;
 }
